vector_bot/requirements.txt
python-dotenv==1.0.1
requests==2.32.3
pydantic==2.7.4
